added test for Dashboard and moved
